Move health server function to health package
